# Clean-up SFC curve sort implementation.