… source/tit/geom/hilbert_ordering.hpp