SFC sorting is cleaned-up and covered with tests. Implementation is parallelized with task groups, and overall shape of the curve has changed slightly.

Performances has increased significantly, due to faster curve sort time (2 seconds), and due to luckier cuts. There may be a room for performance improvement by introducing parallelism into the bounding box computation.

Before:
```
Statistics report:
--------------------------------------------------------------------------------
ParticleAdjacency::color_sizes:
  mean value: 36192 43511 43846 43353 42729 43241 43141 40723 22486
   min value: 32370 41197 41133 39103 38353 40110 40454 29374 12903
   max value: 39726 46439 46565 46455 45132 45203 45635 45374 32075
--------------------------------------------------------------------------------

Profiling report:
--------------------------------------------------------------------------------
abs. time [s]    rel. time [%]    calls [#]    section name
--------------------------------------------------------------------------------
    120.28590        100.00000            1    main
    110.72962         92.05536        34130    EulerIntegrator::step()
     48.74499         40.52428         3413    ParticleAdjacency::build()
     29.00364         24.11226        34130    FluidEquations::compute_density()
     16.72462         13.90406        34130    FluidEquations::setup_boundary()
     14.17562         11.78494        34130    FluidEquations::compute_forces()
      2.93674          2.44146         3413    ZCurveOrdering::ZCurveOrdering()
      0.84294          0.70078         3413    Grid::Grid()
      0.00014          0.00011            1    FluidEquations::init()
--------------------------------------------------------------------------------
```

After:
```
Statistics report:
--------------------------------------------------------------------------------
ParticleAdjacency::color_sizes:
  mean value: 43601 42986 43209 43818 44043 43272 42646 36407 19242
   min value: 38135 39310 40380 41255 40169 38917 37228 26654 9515
   max value: 46289 45389 45399 46188 45643 45169 45271 42419 30152
--------------------------------------------------------------------------------

Profiling report:
--------------------------------------------------------------------------------
abs. time [s]    rel. time [%]    calls [#]    section name
--------------------------------------------------------------------------------
    115.80262        100.00000            1    main
    106.19788         91.70594        34130    EulerIntegrator::step()
     46.28678         39.97041         3413    ParticleAdjacency::build()
     27.34873         23.61667        34130    FluidEquations::compute_density()
     16.77265         14.48383        34130    FluidEquations::setup_boundary()
     13.74169         11.86648        34130    FluidEquations::compute_forces()
      0.84275          0.72775         3413    Grid::Grid()
      0.66117          0.57094         3413    MortonCurveSort::MortonCurveSort()
      0.00013          0.00012            1    FluidEquations::init()
--------------------------------------------------------------------------------
```